Add tags to sharing integration tests to split them in Drone
As the sharing integration tests take very long to run they usually time
out in Drone, so they need to be split and run in smaller sets. To do
this the feature files were tagged as "set-1", "set-2"... and the Drone
pipeline that run all the sharing integration tests was split in several
pipelines, each one running only one of the sets.

The feature files were tagged in alphabetic order and grouping roughly
the same number of scenarios in each set.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>